VAULT: Withdrawal slippage PT1 (#121)
The time between enqueuing withdrawal and burning a slashing event can
happen. This would result in an unfavorable slippage. While this is the
result of slashing, the depositor should be able to decide how much
slippage they are willing to incur and make an informed decision if the
slippage exceeds what they are expecting.

The first part of this solution is to allow the depositor to set the
`min_amount_out` when calling `enqueue_withdrawal`.

[Outline of proposed final
solution](https://www.figma.com/board/jSYab9VYCfDZ7Zp6f0pXYf/Slippage-Fee-Burning-Withdraw-Ticket?node-id=8-220&t=71HAIL90Q9ds4tcE-1)

---------

Co-authored-by: Aoi Kurokawa <62386689+aoikurokawa@users.noreply.github.com>