RESTAKING + VAULT: Adding Metadata Admin Field (#130)
In the future we will have a metadata account for the vaults, operators
and ncns. I think it makes sense to keep the admin to update this
metadata within it's respective account and this keeps our pattern of
separation of authority.

Side effect: This metadata_admin consumed the `reserve` buffer in the
middle of the accounts. Which is good to clean it up a bit