Standardize Token Program Validatation (#175)
All IXs that interact with the Token Program calls `load_token_program`
which checks for the original SPL token id: `spl_token::id()`

There are weird side effects that need to be considered when
implementing Token2022. My suggestion would be to hold off until we have
time to consider the implications of Token2022. When we do, we should
potentially create token_2022 versions of our IXs.