AUDIT: 37314 (#191)
+ If the staker withdrawing is either the program fee wallet or vault
fee wallet, don't take their respective fees

Entry:
37314 (High) "Vault creators can not withdraw their fees without being
recursively charged (vault and program) fees on their own fees which
causes permanent loss of funds"

The program fee is still assessed on withdrawals of VRT from the program
fee wallet, so some amount will always remain in the program fee wallet
for as long as the vault is operational. I do not agree with the
severity. I think this makes more sense as a Medium finding.